[MODEBSNET-3] Moved AsyncFolioExecutionContext to src/test to reach 80% coverage
